LPS-99657 'test.internal' -> 'internal.test'
