+++ modules/apps/oauth2-provider/oauth2-provider-test/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay OAuth2 Provider Test
 Bundle-SymbolicName: com.liferay.oauth2.provider.test
 Bundle-Version: 1.0.0
-Private-Package: com.liferay.oauth2.provider.test.internal.*
+Private-Package: com.liferay.oauth2.provider.internal.test.*
