+++ modules/apps/oauth2-provider/oauth2-provider-test/src/main/java/com/liferay/oauth2/provider/internal/test/TestApplication.java
@@ -12,30 +12,33 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.test.internal;
+package com.liferay.oauth2.provider.internal.test;
 
 import java.util.Collections;
 import java.util.Set;
 
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.core.Application;
 
 /**
  * @author Carlos Sierra Andrés
  */
-public class TestInterfaceAnnotatedApplication
-	extends Application implements TestAnnotatedInterface {
+public class TestApplication extends Application {
 
 	@Override
 	public Set<Object> getSingletons() {
 		return Collections.<Object>singleton(this);
 	}
 
+	@GET
 	public String getString() {
-		return "everything.read";
+		return "get";
 	}
 
-	public String getStringNoScope() {
-		return "no-scope";
+	@POST
+	public String post(String post) {
+		return "post";
 	}
 
 }
